move utils/vc.js into utils/vc/index.js
