+++ src/utils/revocation.js
@@ -20,6 +20,15 @@ export const RevRegIdByteSize = 32;
 export const RevEntryByteSize = 32;
 
 /**
+ * Return `credentialStatus` according to W3C spec when the revocation status is checked on Dock
+ * @param registryId - Revocation registry id
+ * @returns {{id: string, type: string}}
+ */
+export function buildDockCredentialStatus(registryId) {
+  return { id: `${DockRevRegQualifier}${registryId}`, type: RevRegType };
+}
+
+/**
  * Generate the revocation id of a credential usable by Dock. It hashes the credential id to get the
  * revocation id
  * @param credential
